version 8.5.1

+++ module/ASC.Api/ASC.Api.MailServer/MailServerApi.MailGroup.cs
@@ -1,30 +1,28 @@
 /*
- * 
- * (c) Copyright Ascensio System SIA 2010-2014
- * 
- * This program is a free software product.
- * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
- * (AGPL) version 3 as published by the Free Software Foundation. 
- * In accordance with Section 7(a) of the GNU AGPL its Section 15 shall be amended to the effect 
- * that Ascensio System SIA expressly excludes the warranty of non-infringement of any third-party rights.
- * 
- * This program is distributed WITHOUT ANY WARRANTY; 
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
- * For details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
- * 
- * You can contact Ascensio System SIA at Lubanas st. 125a-25, Riga, Latvia, EU, LV-1021.
- * 
- * The interactive user interfaces in modified source and object code versions of the Program 
- * must display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
- * 
- * Pursuant to Section 7(b) of the License you must retain the original Product logo when distributing the program. 
- * Pursuant to Section 7(e) we decline to grant you any rights under trademark law for use of our trademarks.
- * 
- * All the Product's GUI elements, including illustrations and icon sets, as well as technical 
- * writing content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0 International. 
- * See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
- * 
+ *
+ * (c) Copyright Ascensio System Limited 2010-2015
+ *
+ * This program is freeware. You can redistribute it and/or modify it under the terms of the GNU 
+ * General Public License (GPL) version 3 as published by the Free Software Foundation (https://www.gnu.org/copyleft/gpl.html). 
+ * In accordance with Section 7(a) of the GNU GPL its Section 15 shall be amended to the effect that 
+ * Ascensio System SIA expressly excludes the warranty of non-infringement of any third-party rights.
+ *
+ * THIS PROGRAM IS DISTRIBUTED WITHOUT ANY WARRANTY; WITHOUT EVEN THE IMPLIED WARRANTY OF MERCHANTABILITY OR
+ * FITNESS FOR A PARTICULAR PURPOSE. For more details, see GNU GPL at https://www.gnu.org/copyleft/gpl.html
+ *
+ * You can contact Ascensio System SIA by email at sales@onlyoffice.com
+ *
+ * The interactive user interfaces in modified source and object code versions of ONLYOFFICE must display 
+ * Appropriate Legal Notices, as required under Section 5 of the GNU GPL version 3.
+ *
+ * Pursuant to Section 7 § 3(b) of the GNU GPL you must retain the original ONLYOFFICE logo which contains 
+ * relevant author attributions when distributing the software. If the display of the logo in its graphic 
+ * form is not reasonably feasible for technical reasons, you must include the words "Powered by ONLYOFFICE" 
+ * in every copy of the program you distribute. 
+ * Pursuant to Section 7 § 3(e) we decline to grant you any rights under trademark law for use of our trademarks.
+ *
 */
+
 
 using System;
 using System.Collections.Generic;
@@ -32,7 +30,10 @@ using System.Linq;
 using ASC.Api.Attributes;
 using ASC.Api.MailServer.DataContracts;
 using ASC.Api.MailServer.Extensions;
+using ASC.Mail.Aggregator.Common;
 using ASC.Mail.Server.Utils;
+using System.Security;
+using ASC.Web.Studio.Core;
 
 namespace ASC.Api.MailServer
 {
@@ -50,26 +51,32 @@ namespace ASC.Api.MailServer
         [Create(@"groupaddress/add")]
         public MailGroupData CreateMailGroup(string name, int domain_id, List<int> address_ids)
         {
+            if (!IsAdmin)
+                throw new SecurityException("Need admin privileges.");
+
             if (string.IsNullOrEmpty(name))
-                throw new ArgumentException("Invalid mailgroup name.", "name");
+                throw new ArgumentException(@"Invalid mailgroup name.", "name");
 
             if (domain_id < 0)
-                throw new ArgumentException("Invalid domain id.", "domain_id");
+                throw new ArgumentException(@"Invalid domain id.", "domain_id");
 
             if (name.Length > 64)
-                throw new ArgumentException("Local part of mailgroup exceed limitation of 64 characters.", "name");
+                throw new ArgumentException(@"Local part of mailgroup exceed limitation of 64 characters.", "name");
 
             if (!Parser.IsEmailLocalPartValid(name))
-                throw new ArgumentException("Incorrect group name.", "name");
+                throw new ArgumentException(@"Incorrect group name.", "name");
 
             if (!address_ids.Any())
-                throw new ArgumentException("Empty collection of address_ids.", "address_ids");
+                throw new ArgumentException(@"Empty collection of address_ids.", "address_ids");
 
             var domain = MailServer.GetWebDomain(domain_id, MailServerFactory);
 
-            var mailgroup_name = name.ToLowerInvariant();
+            if (domain.Tenant == Defines.SHARED_TENANT_ID)
+                throw new InvalidOperationException("Creating mail group is not allowed for shared domain.");
 
-            var mailgroup = MailServer.CreateMailGroup(mailgroup_name, domain, address_ids, MailServerFactory);
+            var mailgroupName = name.ToLowerInvariant();
+
+            var mailgroup = MailServer.CreateMailGroup(mailgroupName, domain, address_ids, MailServerFactory);
 
             return mailgroup.ToMailGroupData();
         }
@@ -85,13 +92,19 @@ namespace ASC.Api.MailServer
         [Update(@"groupaddress/address/add")]
         public MailGroupData AddMailGroupAddress(int mailgroup_id, int address_id)
         {
+            if (!IsAdmin)
+                throw new SecurityException("Need admin privileges.");
+
             if (address_id < 0)
-                throw new ArgumentException("Invalid address id.", "address_id");
+                throw new ArgumentException(@"Invalid address id.", "address_id");
 
             if (mailgroup_id < 0)
-                throw new ArgumentException("Invalid mailgroup id.", "mailgroup_id");
-            
+                throw new ArgumentException(@"Invalid mailgroup id.", "mailgroup_id");
+
             var mailgroup = MailServer.GetMailGroup(mailgroup_id, MailServerFactory);
+
+            if (mailgroup == null)
+                throw new ArgumentException("Mailgroup not exists");
 
             mailgroup.AddMember(address_id, MailServerFactory);
 
@@ -109,13 +122,19 @@ namespace ASC.Api.MailServer
         [Delete(@"groupaddress/addresses/remove")]
         public int RemoveMailGroupAddress(int mailgroup_id, int address_id)
         {
+            if (!IsAdmin)
+                throw new SecurityException("Need admin privileges.");
+
             if (address_id < 0)
-                throw new ArgumentException("Invalid address id.", "address_id");
+                throw new ArgumentException(@"Invalid address id.", "address_id");
 
             if (mailgroup_id < 0)
-                throw new ArgumentException("Invalid mailgroup id.", "mailgroup_id");
+                throw new ArgumentException(@"Invalid mailgroup id.", "mailgroup_id");
             
             var mailgroup = MailServer.GetMailGroup(mailgroup_id, MailServerFactory);
+
+            if (mailgroup == null)
+                throw new ArgumentException("Mailgroup not exists");
 
             mailgroup.RemoveMember(address_id);
 
@@ -131,6 +150,9 @@ namespace ASC.Api.MailServer
         [Read(@"groupaddress/get")]
         public List<MailGroupData> GetMailGroups()
         {
+            if (!IsAdmin)
+                throw new SecurityException("Need admin privileges.");
+
             var groups = MailServer.GetMailGroups(MailServerFactory);
 
             return groups
@@ -148,8 +170,11 @@ namespace ASC.Api.MailServer
         [Delete(@"groupaddress/remove/{id}")]
         public int RemoveMailGroup(int id)
         {
+            if (!IsAdmin)
+                throw new SecurityException("Need admin privileges.");
+
             if (id < 0)
-                throw new ArgumentException("Invalid mailgroup id.", "id");
+                throw new ArgumentException(@"Invalid mailgroup id.", "id");
 
             MailServer.DeleteMailGroup(id, MailServerFactory);
